Refactor ProcessingDialog
