+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/tvhome/components/TvScreenMenu.kt
@@ -381,7 +381,7 @@ fun TvScreenMenu(
                     Text(text = "Auto play channel")
                 }
 
-                if (true) {
+                if (false) {
                     if (startTvAutomatically) {
                         Spacer(modifier = Modifier.height(8.dp))
                         Row(
